+++ Views/Admin/components/NavigationList.vue
@@ -22,10 +22,10 @@
             <ol class="breadcrumb">
                 <li class="active">Menu</li>
             </ol>
-            <router-link class="btn ink-reaction btn-raised btn-lg btn-info pull-right" :to="{name: 'module:navigation:action', params: {website_id: website_id, navigation_id: 'create'}}">
+            <router-link v-if="auth.status.level < 4" class="btn ink-reaction btn-raised btn-lg btn-info pull-right" :to="{name: 'module:navigation:action', params: {website_id: website_id, navigation_id: 'create'}}">
                 <i class="fa fa-plus"></i> Ajouter un menu
             </router-link>
-            <div class="btn-group pull-right">
+            <div v-if="auth.status.level < 4" class="btn-group pull-right">
                 <button type="button" class="btn btn-lg ink-reaction btn-primary">Action</button>
                 <button type="button" class="btn btn-lg ink-reaction btn-primary dropdown-toggle" data-toggle="dropdown" aria-expanded="false"><i class="fa fa-caret-down"></i></button>
                 <ul class="dropdown-menu dropdown-menu-right" role="menu">
@@ -48,7 +48,7 @@
                                 <table class="table no-margin">
                                     <thead>
                                         <tr>
-                                            <th>
+                                            <th v-show="auth.status.level < 4">
                                                 <div class="checkbox checkbox-styled">
                                                     <label>
                                                         <input v-model="selectAll" type="checkbox">
@@ -66,7 +66,7 @@
                                             <td colspan="4">Aucuns menus trouvés</td>
                                         </tr>
                                         <tr v-for="navigation in navigations">
-                                            <td>
+                                            <td v-show="auth.status.level < 4">
                                                 <div class="checkbox checkbox-styled">
                                                     <label>
                                                         <input type="checkbox" :value="navigation.id"
@@ -84,13 +84,13 @@
                                             <td>
                                                 <router-link
                                                         :to="{name: 'module:navigation:action', params: {website_id: website_id, navigation_id: navigation.id}}"
-                                                        class="btn ink-reaction btn-floating-action btn-info">
-                                                    <i class="fa fa-pencil"></i>
+                                                        class="btn ink-reaction btn-info">
+                                                    <i class="fa fa-pencil"></i> Modifier
                                                 </router-link>
-                                                <a @click="selectNavigation(navigation.id)" data-toggle="modal"
+                                                <a v-if="auth.status.level < 4" @click="selectNavigation(navigation.id)" data-toggle="modal"
                                                    data-target="#deleteNavigationModal"
-                                                   class="btn ink-reaction btn-floating-action btn-danger"><i
-                                                        class="fa fa-trash"></i></a>
+                                                   class="btn ink-reaction btn-danger"><i
+                                                        class="fa fa-trash"></i> Supprimer</a>
                                             </td>
                                         </tr>
                                     </tbody>
@@ -170,16 +170,18 @@
                 this.selected_items = [id];
             },
             deleteNavigation(){
-                this.destroy({
-                    api: navigation_api.destroy + this.website_id,
-                    ids: this.selected_items
-                }).then(() => {
-                    this.selected_items.forEach((el) => {
-                        let index = this.navigations.findIndex((nav) => nav.id == el);
-                        this.navigations.splice(index,1);
-                    })
-                    this.selected_items = [];
-                });
+                if(this.auth.status.level < 4){
+                    this.destroy({
+                        api: navigation_api.destroy + this.website_id,
+                        ids: this.selected_items
+                    }).then(() => {
+                        this.selected_items.forEach((el) => {
+                            let index = this.navigations.findIndex((nav) => nav.id == el);
+                            this.navigations.splice(index,1);
+                        })
+                        this.selected_items = [];
+                    });
+                }
             }
         },
         created () {
